requirements.txt
Django
djangorestframework
django-mysql